+++ simbak/agent/__init__.py
@@ -1,7 +1,9 @@
 from simbak.agent.base import BaseAgent
 from simbak.agent.normal import NormalAgent
+from simbak.agent.rotating import RotatingAgent
 
 __all__ = [
     'BaseAgent',
     'NormalAgent',
+    'RotatingAgent',
 ]