+++ simbak/agent/normal.py
@@ -1,7 +1,5 @@
 import logging as _logging
-from typing import Tuple as _Tuple
 
-from simbak import fileutil as _fileutil
 from simbak.agent.base import BaseAgent as _BaseAgent
 
 _logger = _logging.getLogger(__name__)
@@ -10,6 +8,17 @@ _logger = _logging.getLogger(__name__)
 class NormalAgent(_BaseAgent):
     def __init__(self, sources: list, destinations: list, name: str,
                  compression_level: int = 6):
+        """
+        Args:
+            sources (list of str): Paths to the files that you are
+                backing up.
+            destinations (list of str): Paths of where you want the
+                backup to be stored.
+            name (str): Name of the backup, this will name the backup
+                files.
+            compression_level (int, optional): The gzip compression
+                level that you want to use for the backup. Default to 6.
+        """
         super().__init__(sources, destinations, name, compression_level)
 
     def backup(self):
@@ -20,34 +29,5 @@ class NormalAgent(_BaseAgent):
         the gzip file will be the name parameter suffixed with a time
         stamp, the format of the timestamp is YYYY-MM-DD--hh-mm-ss
         """
-        # Cleans the path names, and removes non-existent paths, if relevant.
         _logger.info(f'Starting backup [{self._name}]')
-        sources, destinations = self._filter_paths()
-        super()._validate_paths(sources, destinations)
-
-        # A unique file name is important for certain backup agents.
-        file_name = _fileutil.unique_file_name(self._name)
-        _logger.info(f'Backup file name will be {self._name}')
-
-        # Using tar gzip for the backup. Tarring once, distributing later.
-        first_path = _fileutil.create_targz(
-            sources=sources,
-            destination=destinations[0],
-            file_name=file_name,
-            compression_level=self._compression_level
-        )
-
-        super()._log_source_sizes(sources)
-        super()._log_backup_size(first_path)
-
-        _fileutil.distribute_file(
-            path=first_path,
-            destinations=destinations[1:]
-        )
-
-    def _filter_paths(self, create: bool = True) -> _Tuple:
-        sources = _fileutil.filter_paths(self._sources)
-        destinations = _fileutil.filter_paths(
-            self._destinations, create=create)
-
-        return sources, destinations
+        super()._backup()